+++ NEWS.md
@@ -1,3 +1,8 @@
+# a11ytables 0.0.0.9003
+
+* BREAKING CHANGE: the `add_*()` function family for building a workbook are retired, given the introduction of `create_a11y_wb()`
+* Introduced S3 class 'a11ytable', including exported functions to generate (`new_a11ytable()`, `as_a11ytable`) and test (`is_a11ytable`) objects of this class
+
 # a11ytables 0.0.0.9002
 
 * Simplify all `add_*()` functions to a single `create_a11y_wb()` function (moving towards #15), though `add_*()` functions are all still exported